+++ microservices/ol-service/src/main/java/com/github/valiclud/old_languages/core/ol/services/OldLanguageServiceImpl.java
@@ -22,7 +22,7 @@ public class OldLanguageServiceImpl implements OldLanguageService {
 	}
 
 	@Override
-	public OldLanguage getOldLanguage(Long oldLanguageId) {
+	public OldLanguage getOldLanguage(int oldLanguageId) {
 		LOG.debug("/product return the found product for oldLaguageId={}", oldLanguageId);
 
 		if (oldLanguageId < 1) {
